fixes Retrofit conflicts
