+++ app/src/main/java/com/melvin/ongandroid/businesslogic/domain/ApiService.kt
@@ -2,7 +2,9 @@ package com.melvin.ongandroid.businesslogic.domain
 
 import com.melvin.ongandroid.model.DefaultResponse
 import com.melvin.ongandroid.model.User
+import com.melvin.ongandroid.model.response.VerifyUser
 import retrofit2.Call
+import retrofit2.Response
 import retrofit2.http.*
 
 
@@ -12,4 +14,10 @@ interface ApiService {
     fun createUser(
        @Body user: User
     ): Call<DefaultResponse>
+
+    @POST("api/login")
+    suspend fun postLogin(
+        @Query("email") user: String,
+        @Query("password") pass: String
+    ): Response<VerifyUser>
 }
